WIP Porting: Move to Declarators Collector
 - Implemented Info and Declarative objects for Shapes, Sequences,
Configs, Contexts
 - Implemented Info object for Pulse (later might delete it and
directly move to a sole ItemInfos)
 - converted the manifest and plugin to handle those through
Declarations in the manifest
 - Added code to handle collection of filters in the plugin
 - Renamed the sequences_request and shapes_request to
get_sequences_infos, get_shapes_infos

 - Modified the way templates are requested and stored: when a
new/changed ini file is detected, it’s path get stored in a
SequenceInfo object. When this is requested, the actual data in the ini
is read. Note: Current storage with redundant template_sequence_data
and _infos is redundant. Will be changed